+++ nudebomb/walk.py
@@ -4,6 +4,7 @@ import os
 from copy import deepcopy
 from pathlib import Path
 
+from termcolor import cprint
 from treestamps import Treestamps
 
 from nudebomb.config import TIMESTAMPS_CONFIG_KEYS
@@ -18,7 +19,7 @@ class Walk:
     def __init__(self, config):
         """Initialize."""
         self._config = config
-        self._langfiles = LangFiles(config.languages)
+        self._langfiles = LangFiles(config)
         self._timestamps: dict[Path, Treestamps] = {}
 
     def _is_path_ignored(self, path: Path) -> bool:
@@ -40,10 +41,13 @@ class Walk:
             mtime = self._timestamps.get(top_path, {}).get(path)
 
         if mtime is not None and mtime > path.stat().st_mtime:
+            if self._config.verbose:
+                cprint(f"Skip unchanged {path}", "white", attrs=["dark"])
             return
 
+        dirpath = Treestamps.dirpath(path)
         config = deepcopy(self._config)
-        config.languages = self._langfiles.get_langs(top_path, path)
+        config.languages = self._langfiles.get_langs(top_path, dirpath)
         mkv_obj = MKVFile(config, path)
         mkv_obj.remove_tracks()
 
@@ -55,42 +59,55 @@ class Walk:
         if not self._config.recurse:
             return
 
-        dirs = []
         filenames = []
 
-        for filename in os.scandir(dir):
-            path = Path(filename)
-            if path.is_dir():
-                dirs.append(path)
+        for filename in sorted(os.listdir(dir)):
+            entry_path = dir / filename
+            if entry_path.is_dir():
+                self.walk_file(top_path, entry_path)
             else:
-                filenames.append(path)
+                filenames.append(entry_path)
 
-        for path in sorted(dirs):
-            self.walk_dir(top_path, path)
-
-        for path in sorted(filenames):
+        for path in filenames:
             self.walk_file(top_path, path)
 
         if self._config.timestamps:
             timestamps = self._timestamps[top_path]
             timestamps.set(dir, compact=True)
-            timestamps.dump()
 
     def walk_file(self, top_path, path):
         """Walk a file."""
-        if self._is_path_ignored(path) or (
-            not self._config.symlinks and path.is_symlink()
-        ):
+        if self._is_path_ignored(path):
+            if self._config.verbose:
+                cprint(f"Skip ignored {path}", "white", attrs=["dark"])
+            return
+        if not self._config.symlinks and path.is_symlink():
+            if self._config.verbose:
+                cprint(f"Skip symlink {path}", "white", attrs=["dark"])
             return
         if path.is_dir():
             self.walk_dir(top_path, path)
         else:
             self.strip_path(top_path, path)
 
-    def run(self):
-        """Run the stripper against all configured paths."""
+    def print_info(self):
+        """Print intentions before we begin."""
+        langs = ", ".join(sorted(self._config.languages))
+        if self._config.sub_languages:
+            audio = "audio "
+        else:
+            audio = ""
+        print(f"Stripping {audio}languages except {langs}.")
+        if self._config.sub_languages:
+            sub_langs = ", ".join(sorted(self._config.sub_languages))
+            print(f"Stripping subtitle languages except {sub_langs}.")
+
         if self._config.verbose:
             print("Searching for MKV files to process...")
+
+    def run(self):
+        """Run the stripper against all configured paths."""
+        self.print_info()
 
         if self._config.timestamps:
             self._timestamps = Treestamps.map_factory(
@@ -108,5 +125,6 @@ class Walk:
             self.walk_file(top_path, path)
 
         if self._config.timestamps:
-            for timestamps in self._timestamps.values():
+            for top_path, timestamps in self._timestamps.items():
+                print(f"Saving timestamps for {top_path}")
                 timestamps.dump()