+++ NEWS.md
@@ -1,5 +1,16 @@
 # 📰 Nudebomb News
 
+## v0.2.3
+
+- Features
+
+  - More output about what nudebomb is doing.
+  - Renamed subs_languages config option to sub_languages.
+
+- Fixes
+
+  - Searching for langfiles obeys symlinks preference.
+
 ## v0.2.2
 
 - Fixes